Fixed a bug with a function loop on startup
